backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai==1.30.5
pydantic==2.7.1
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.9
httpx==0.27.0
